requirements.txt
# 基于RAG的游戏问答系统依赖包列表

# ==================== 核心RAG框架 ====================
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
qdrant-client==1.7.0

# ==================== 大语言模型 ====================
transformers==4.35.0
torch==2.1.0
openai==1.3.0
accelerate==0.24.0
bitsandbytes==0.41.0

# ==================== 数据处理 ====================
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
jieba==0.42.1
pypinyin==0.49.0

# ==================== 多模态交互 ====================
# 语音处理
speechrecognition==3.10.0
pyttsx3==2.90
librosa==0.10.1
soundfile==0.12.1

# 视觉处理
opencv-python==4.8.0
pillow==10.0.0
torchvision==0.16.0
albumentations==1.3.1

# 触觉反馈
pygame==2.5.0

# ==================== Web框架 ====================
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.4.0
python-multipart==0.0.6
aiofiles==23.2.1

# ==================== 数据库 ====================
sqlalchemy==2.0.0
alembic==1.12.0
redis==5.0.0
psycopg2-binary==2.9.0
asyncpg==0.29.0

# ==================== 爬虫系统 ====================
scrapy==2.11.0
aiohttp==3.9.0
requests==2.31.0
beautifulsoup4==4.12.0
selenium==4.15.0
playwright==1.40.0

# ==================== 向量存储 ====================
chromadb==0.4.15
pinecone-client==2.2.4
weaviate-client==3.25.0

# ==================== 可视化分析 ====================
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.17.0
streamlit==1.28.0
gradio==4.0.0

# ==================== 无障碍支持 ====================
# 文本处理
pypinyin==0.49.0
cn2an==6.2.0

# 图像生成
matplotlib==3.7.0
pillow==10.0.0

# ==================== 健康管理 ====================
# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# 数据序列化
msgpack==1.0.7
orjson==3.9.10

# ==================== 工具库 ====================
python-dotenv==1.0.0
loguru==0.7.0
rich==13.7.0
click==8.1.7
typer==0.9.0

# ==================== 测试框架 ====================
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0

# ==================== 部署相关 ====================
gunicorn==21.2.0
docker==6.1.3
kubernetes==28.1.0

# ==================== 开发工具 ====================
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0

# ==================== 监控日志 ====================
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# ==================== 安全认证 ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ==================== 性能优化 ====================
uvloop==0.19.0
orjson==3.9.10
ujson==5.8.0

# ==================== 可选依赖 ====================
# GPU支持 (取消注释以启用)
# torch-audio==2.1.0
# torch-vision==0.16.0

# 高级NLP (取消注释以启用)
# spacy==3.7.0
# nltk==3.8.1

# 机器学习扩展 (取消注释以启用)
# xgboost==2.0.2
# lightgbm==4.1.0
# catboost==1.2.2
